Added Germy Demo Level
Alex created our demo level concept, which is our goal for the next two weeks. I scaled it to fit and imported it into the master file. I've also fixed some of Germy Jack's AI in preperation for the meeting, and solved any prefab errors I could find at the moment. In short, this commit is a lot of updates but should be the one with the best logic so far.