Morning bug patch
Deleted outdated versions of maps, fixed a few issues with the playerRB script and the boostshoes script to make them closer to what we want. More functinality is needed but it's starting to look excellent.

Also applied all current changes to the Prefab, so just drag and drop a PlayerRB to test him out